Jump to message from notification now working
